fix(dashboard): try fix dashboard serialization bug on timestamp

+++ kun-data-dashboard/src/main/java/com/miotech/kun/datadashboard/model/entity/DateTimeTaskCount.java
@@ -9,7 +9,8 @@ import lombok.Data;
 @Data
 public class DateTimeTaskCount {
 
-    Long time;
+    /* Epoch Unix timestamp in milliseconds */
+    Double time;
 
     Integer taskCount;
 }
